refactor 3d building layer block to own file
